fixed dashbord issue and renamed files and refactored with feedback
