+++ src/client/components/chart-data-buttons/chartbar-buttons/chart-data-button.stories.js
@@ -19,5 +19,5 @@ const defaultValue = [
 
 export const ChartdataButtons = () => {
   const value = object(label, defaultValue);
-  return <ChartbarMenu buttons={value} defaultSelection={{ id: 2 }} />;
+  return <ChartbarMenu buttons={value} />;
 };
