+++ src/client/components/dashboard-items/dashboard-items.component.js
@@ -2,11 +2,10 @@ import React, { useContext } from 'react';
 import './dashboard-items.style.css';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { ChartDataContext } from '../../containers/chart-detail-page/chart-detail-context';
+import { ChartDataContext } from '../../containers/chart-detail-page/chart-detail-page.context';
 
 const DashboardItems = ({ items }) => {
-  const listItemClicked = useContext(ChartDataContext);
-  const { handleClick } = listItemClicked;
+  const { handleClick } = useContext(ChartDataContext);
   return (
     <ul className="list">
       {items.map((item) => (
