+++ src/client/components/side-navigation/sidebar.stories.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import SidebarMenu from './sidebar.component';
-import { action } from '@storybook/addon-actions';
 import { withKnobs } from '@storybook/addon-knobs';
 import '@storybook/addon-knobs/register';
 
@@ -11,7 +10,5 @@ export default {
 };
 
 export const SidebarActive = () => {
-  const text = ('value', 'dashboard');
-  const handleClick = action('clicked');
-  return <SidebarMenu text={text} handleClick={handleClick} />;
+  return <SidebarMenu />;
 };
