+++ src/client/containers/chart-detail-page/chart-detail-page.css
